+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/src/runtime/adapter.ts
@@ -12,9 +12,12 @@ import { normaliseBase as canonicalNormaliseBase } from "./utils/normaliseBase.j
 import { createExclude, type ExcludeConfig } from "./exclude.js"
 import { guessTextDirMap } from "./utils/text-dir.js"
 import { resolvePathTranslations } from "./config/resolvePathTranslations.js"
-import { validatePathTranslations } from "./config/validatePathTranslations.js"
+import {
+	validatePathTranslations,
+	type PathDefinitionTranslations,
+} from "@inlang/paraglide-js/internal/adapter-utils"
 import type { ParamMatcher } from "@sveltejs/kit"
-import type { PathTranslations, UserPathTranslations } from "./config/pathTranslations.js"
+import type { UserPathTranslations } from "./config/pathTranslations.js"
 import type { Paraglide } from "./runtime.js"
 
 export type I18nUserConfig<T extends string> = {
@@ -121,7 +124,7 @@ export type I18nUserConfig<T extends string> = {
  */
 export type I18nConfig<T extends string> = {
 	runtime: Paraglide<T>
-	translations: PathTranslations<T>
+	translations: PathDefinitionTranslations<T>
 	exclude: (path: string) => boolean
 	defaultLanguageTag: T
 	matchers: Record<string, ParamMatcher>
@@ -151,18 +154,18 @@ export type I18nConfig<T extends string> = {
 export function createI18n<T extends string>(runtime: Paraglide<T>, options?: I18nUserConfig<T>) {
 	const translations = resolvePathTranslations(
 		options?.pathnames ?? {},
-		runtime.availableLanguageTags
+		runtime.availableLanguageTags,
 	)
 
 	if (dev) {
 		const issues = validatePathTranslations(
 			translations,
 			runtime.availableLanguageTags,
-			options?.matchers ?? {}
+			options?.matchers ?? {},
 		)
 		if (issues.length) {
 			console.warn(
-				`The following issues were found in your path translations. Make sure to fix them before deploying your app:`
+				`The following issues were found in your path translations. Make sure to fix them before deploying your app:`,
 			)
 			console.table(issues)
 		}
@@ -270,7 +273,7 @@ export function createI18n<T extends string>(runtime: Paraglide<T>, options?: I1
 				canonicalPath,
 				lang,
 				config.translations,
-				config.matchers
+				config.matchers,
 			)
 
 			return serializeRoute({
@@ -304,7 +307,6 @@ export function createI18n<T extends string>(runtime: Paraglide<T>, options?: I1
 		 * ```
 		 */
 		route(translatedPath: string) {
-
 			const normalizedBase = normaliseBase(base)
 
 			const { path, lang, trailingSlash, dataSuffix } = getPathInfo(translatedPath, {