+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/src/runtime/path-translations/getTranslatedPath.ts
@@ -1,12 +1,14 @@
-import { bestMatch } from "./matching/match.js"
-import { resolve_route } from "./matching/routing.js"
-import type { PathTranslations } from "../config/pathTranslations.js"
+import {
+	resolveRoute,
+	bestMatch,
+	type PathDefinitionTranslations,
+} from "@inlang/paraglide-js/internal/adapter-utils"
 import type { ParamMatcher } from "@sveltejs/kit"
 
 export function getTranslatedPath(
 	canonicalPath: string,
 	lang: string,
-	translations: PathTranslations,
+	translations: PathDefinitionTranslations,
 	matchers: Record<string, ParamMatcher>
 ) {
 	const match = bestMatch(canonicalPath, Object.keys(translations), matchers)
@@ -18,5 +20,5 @@ export function getTranslatedPath(
 	const translatedPath = translationsForPath[lang]
 	if (!translatedPath) return canonicalPath
 
-	return resolve_route(translatedPath, match.params)
+	return resolveRoute(translatedPath, match.params)
 }