+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/turbo/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @inlang/paraglide-js-adapter-next-example-turbo
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [fa6aa31]
+- Updated dependencies [dee5aa6]
+  - @inlang/paraglide-js@1.6.1
+
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies [462325b]
+  - @inlang/paraglide-js@1.6.0
+
 ## 0.1.14
 
 ### Patch Changes